# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synque Express Ticketing API",
    "version": "0.1.0",
    "description": "Complete ticketing and event management API with user profile management, promotions, and cancellation support"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    },
    {
      "url": "https://api.ticketing.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/payments/wechat/prepay": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create WeChat prepay order via Wallyt",
        "description": "Create a prepay order for WeChat Mini Program payment. Returns payment parameters for wx.requestPayment().",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId", "openid"],
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "Order ID"
                  },
                  "openid": {
                    "type": "string",
                    "description": "User's OpenID in the Mini Program"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prepay order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WallytPrepayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Order status does not allow payment"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          },
          "422": {
            "description": "Invalid request parameters"
          },
          "503": {
            "description": "Payment service not configured"
          }
        }
      }
    },
    "/payments/wallyt/notify": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Wallyt payment callback",
        "description": "Wallyt server callback for payment notification. XML format request body. Must return 'success' string within 5 seconds.",
        "requestBody": {
          "required": true,
          "content": {
            "application/xml": {
              "schema": {
                "type": "string",
                "description": "XML formatted notification from Wallyt"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns 'success' or 'fail' as plain text",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "enum": ["success", "fail"]
                }
              }
            }
          }
        }
      }
    },
    "/payments/status/{orderId}": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Query payment status",
        "description": "Query the payment status of an order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Order ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/payments/refund": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Apply for refund",
        "description": "Apply for a full or partial refund for a paid order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["orderId"],
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "Order ID"
                  },
                  "refundAmount": {
                    "type": "number",
                    "description": "Refund amount in CNY. If not provided, full refund will be applied."
                  },
                  "reason": {
                    "type": "string",
                    "description": "Reason for refund"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Order status does not allow refund or refund amount exceeds limit"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Order or payment record not found"
          },
          "500": {
            "description": "Refund processing failed"
          },
          "503": {
            "description": "Payment service not configured"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "WallytPrepayResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "orderId": {
                "type": "integer",
                "description": "Order ID"
              },
              "outTradeNo": {
                "type": "string",
                "description": "Merchant order number sent to Wallyt"
              },
              "payParams": {
                "type": "object",
                "description": "Parameters for wx.requestPayment()",
                "properties": {
                  "appId": {
                    "type": "string"
                  },
                  "timeStamp": {
                    "type": "string"
                  },
                  "nonceStr": {
                    "type": "string"
                  },
                  "package": {
                    "type": "string"
                  },
                  "signType": {
                    "type": "string"
                  },
                  "paySign": {
                    "type": "string"
                  }
                }
              },
              "paymentId": {
                "type": "integer",
                "description": "Payment record ID"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "description": "Prepay expiration time"
              }
            }
          }
        }
      },
      "PaymentStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": ["PENDING", "CONFIRMED", "CANCELLED", "REFUNDED"],
                "description": "Order payment status"
              },
              "paidAt": {
                "type": "string",
                "format": "date-time",
                "description": "Payment completion time"
              },
              "transactionId": {
                "type": "string",
                "description": "Wallyt transaction ID"
              }
            }
          }
        }
      },
      "RefundResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "orderId": {
                "type": "integer",
                "description": "Order ID"
              },
              "outRefundNo": {
                "type": "string",
                "description": "Merchant refund order number"
              },
              "refundId": {
                "type": "string",
                "description": "Wallyt refund ID"
              },
              "refundAmount": {
                "type": "number",
                "description": "Refund amount in CNY"
              },
              "status": {
                "type": "string",
                "enum": ["SUCCESS", "PROCESSING", "FAILED"],
                "description": "Refund status"
              },
              "errorMessage": {
                "type": "string",
                "description": "Error message if refund failed"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated endpoints"
      },
      "otaApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for OTA platform integration"
      }
    }
  },
...
}